+++ back/src/test/java/fr/centralesupelec/thuv/test_connection_scheduler/TestSocketTest.java
@@ -9,19 +9,20 @@ import static org.junit.jupiter.api.Assertions.*;
 
 class TestSocketTest {
     TestSocket testSocket;
+    Integer timeout = 3;
 
     @Test
     void isHttpsAliveWithVerification() {
         testSocket = new TestSocket();
         ReflectionTestUtils.setField(testSocket, "disableSslValidation", false);
         ReflectionTestUtils.setField(testSocket, "errorCodes", List.of(404, 410, 503));
-        assertTrue(testSocket.isHttpsAlive("badssl.com", 1));
-        assertTrue(testSocket.isHttpsAlive("httpstat.us/401", 1));
-        assertFalse(testSocket.isHttpsAlive("httpstat.us/404", 1));
-        assertFalse(testSocket.isHttpsAlive("httpstat.us/410", 1));
-        assertFalse(testSocket.isHttpsAlive("httpstat.us/503", 1));
-        assertFalse(testSocket.isHttpsAlive("expired.badssl.com", 1));
-        assertFalse(testSocket.isHttpsAlive("wrong.host.badssl.com", 1));
+        assertTrue(testSocket.isHttpsAlive("badssl.com", timeout));
+        assertTrue(testSocket.isHttpsAlive("httpstat.us/401", timeout));
+        assertFalse(testSocket.isHttpsAlive("httpstat.us/404", timeout));
+        assertFalse(testSocket.isHttpsAlive("httpstat.us/410", timeout));
+        assertFalse(testSocket.isHttpsAlive("httpstat.us/503", timeout));
+        assertFalse(testSocket.isHttpsAlive("expired.badssl.com", timeout));
+        assertFalse(testSocket.isHttpsAlive("wrong.host.badssl.com", timeout));
     }
 
     @Test
@@ -29,11 +30,11 @@ class TestSocketTest {
         testSocket = new TestSocket();
         ReflectionTestUtils.setField(testSocket, "errorCodes", List.of(410, 503));
         ReflectionTestUtils.setField(testSocket, "disableSslValidation", true);
-        assertTrue(testSocket.isHttpsAlive("badssl.com", 1));
-        assertTrue(testSocket.isHttpsAlive("expired.badssl.com", 1));
-        assertTrue(testSocket.isHttpsAlive("wrong.host.badssl.com", 1));
-        assertTrue(testSocket.isHttpsAlive("httpstat.us/404", 1));
-        assertFalse(testSocket.isHttpsAlive("httpstat.us/410", 1));
-        assertFalse(testSocket.isHttpsAlive("httpstat.us/503", 1));
+        assertTrue(testSocket.isHttpsAlive("badssl.com", timeout));
+        assertTrue(testSocket.isHttpsAlive("expired.badssl.com", timeout));
+        assertTrue(testSocket.isHttpsAlive("wrong.host.badssl.com", timeout));
+        assertTrue(testSocket.isHttpsAlive("httpstat.us/404", timeout));
+        assertFalse(testSocket.isHttpsAlive("httpstat.us/410", timeout));
+        assertFalse(testSocket.isHttpsAlive("httpstat.us/503", timeout));
     }
 }