+++ back/CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## 2.31.0
 ### Added
 - Add container swarm state retrieval
 - Add docker environment to deploy / dev the application